+++ defaults.py
@@ -306,6 +306,10 @@ class FlatCAMDefaults:
 
         # Geometry Options
         "tools_mill_tooldia": "2.4",
+        "tools_mill_offset_type": 0,    # _('Path')
+        "tools_mill_offset": 0.0,
+        "tools_mill_job_type": 0,   # 'Rough'
+
         "tools_mill_cutz": -2.4,
         "tools_mill_vtipdia": 0.1,
         "tools_mill_vtipangle": 30,